Update CItyBikeLanes
Which locations in New York received at least 3 orders in January, and how many orders did they each receive?
How many of each type of headphone were sold in February?
What was the average amount spent per account in February? 
What was the average quantity of products purchased per account in February? 
Which product brought in the most revenue in January and how much revenue did it bring in total?